rearange the sr policy code
